Components and style fixes. Closed #18.
Added subtitle for SumAndProduct page.
Fixed the appearance and styles of components on the SumAndProduct page. Added processing for the case when periods are not relatively prime; now, in addition to the usual signature, a modal window is displayed. Fixed a bug with incorrect display of a pseudo-random sequence sequence block width. Changed the names of components and page paths.